Using repoman to merge the reposync downloaded repos
That is a first step towards allowing other rpm sources (anything
supported by repoman, from local dirs, to jenkins or koji builds)
